Add tests for toggle_visibility. Relates to #5
