+++ examples/toggle_visibility/index.tsx
@@ -1,7 +1,8 @@
-import xs, { Stream } from 'xstream'
+import { Stream } from 'xstream'
 import { run, Drivers, FantasyObservable } from '@cycle/run'
 
-import { h, makeDOMDriver, DOMSource } from '../../src/cycle'
+import { makeDOMDriver, DOMSource } from '../../src/cycle'
+import { main } from './main'
 
 export interface Sources {
   DOM: DOMSource,
@@ -15,22 +16,6 @@ export interface Sinks {
 
 const drivers: Drivers<Sources, Sinks> = {
   DOM: makeDOMDriver('#app'),
-}
-
-function main({ DOM }: Sources): Sinks {
-  const visible$ = DOM.selectEvents('.toggle', 'click')
-    .fold(visible => !visible, false)
-
-  const vtree = (
-    <div>
-      <button class='toggle'>toggle</button>
-      <div if$={visible$}>content</div>
-    </div>
-  )
-
-  return {
-    DOM: xs.of(vtree),
-  }
 }
 
 run(main, drivers)
